% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilimod.R
\name{dilimod}
\alias{dilimod}
\title{Bayesian dili proportional odds model with Stan}
\usage{
dilimod(standata, seedn = 123, ...)
}
\arguments{
\item{standata}{list containing input data N, P, y, X, sigma_prior, make_pred, N_pred, X_pred}

\item{seedn}{Seed to use to get reproducible results}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
A list of posterior samples returned by \code{rstan::extract}
}
\description{
Bayesian dili proportional odds model with Stan
}
