% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstandilimod-package.R
\docType{package}
\name{rstandilimod-package}
\alias{rstandilimod-package}
\alias{rstandilimod}
\title{The 'rstandilimod' package.}
\description{
A minimal rstan package with compiled dili mod
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
